% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_recipients_to_excel.R
\name{export_recipients_to_excel}
\alias{export_recipients_to_excel}
\title{export_recipients_to_excel}
\usage{
export_recipients_to_excel(
  stratification,
  file = "recipients.xlsx",
  overwrite = FALSE
)
}
\arguments{
\item{stratification}{A vector of things you want to stratify by. It can be
length of one if you don't really have any stratifications.}

\item{file}{The name of the exported excel file.}

\item{overwrite}{Overwrite an existing recipients.xlsx or not.
Default is FALSE.}
}
\value{
TRUE if the file is created, else FALSE.
}
\description{
Creates an Excel file, called recipients by default, which can be used to
assign recipients to specific stratification reports. You can also share this
file with PI/Outcomes/Product to determine who should receive what.
}
\details{
This file will contain a single worksheet. This worksheet will have four
columns:
\enumerate{
\item stratification: The name of the stratification. This could be, for
example, a site name.
\item to: One or more email addresses to send the report to. Email addresses
should be separated by a comma. ALL stratifications MUST have an email
address, even if it is to a dummy account.
\item cc: One or more email addresses to cc. May be blank.
\item bcc: One or more email addresses to bcc. May be blank.
}

As exported, the only column with information will be the stratification
column.
}
