% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_vocally.R
\name{fail_vocally}
\alias{fail_vocally}
\title{fail_vocally}
\usage{
fail_vocally(err = NULL)
}
\arguments{
\item{err}{The error you wish to handle.}
}
\value{
NOTHING! This thing either creates a message and/or pushes the error
to Slack.
}
\description{
Pushes errors to the best possible destination. If we are running the
rsconnect config, it tries to publish errors to Slack. Otherwise, it pushes
out messages.
}
