% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_strat_col.R
\name{find_strat_col}
\alias{find_strat_col}
\title{find_strat_col}
\usage{
find_strat_col(data = NULL, col_name = "Stratification")
}
\arguments{
\item{data}{The data frame to search through.}

\item{col_name}{The column name value of the stratification column. By
default, this is 'Stratification'. It can, however, be anything else you
want it to be.}
}
\value{
This function will return either a number or NA if it cannot locate a
stratification column.
}
\description{
From a data frame/tibble, find the column labeled "Stratification".
This is the column we will use to stratify the reports by.
If there isn't a stratification column, find_strat_col returns NA.
}
