% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_recipients_from_excel.R
\name{import_recipients_from_excel}
\alias{import_recipients_from_excel}
\title{import_recipients_from_excel}
\usage{
import_recipients_from_excel(
  file = "recipients.xlsx",
  configuration = "default",
  preview = FALSE
)
}
\arguments{
\item{file}{The recipients xlsx file. Default value is 'recipients.xlsx'.}

\item{configuration}{Which named configuration. Values must be "default",
"prod", or "rsconnect".}

\item{preview}{If TRUE, the results are NOT written to config.yml. Instead,
they are previewed by printing. This is only useful interactively. Don't use
this in an automated script, except for amusement purposes.}
}
\value{
TRUE if successful, else FALSE.
}
\description{
Imports recipients for a report from Excel. This makes your life suck less.
This file must contain a single worksheet. This worksheet shall have four
columns:
\enumerate{
\item stratification: The name of the stratification. This could be, for
example, a site name.
\item to: One or more email addresses to send the report to. Email addresses
should be separated by a comma. ALL stratifications MUST have an email
address, even if it is to a dummy account.
\item cc: One or more email addresses to cc. May be blank.
\item bcc: One or more email addresses to bcc. May be blank.
}
}
