% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_config_recipients.R
\name{reset_config_recipients}
\alias{reset_config_recipients}
\title{reset_config_recipients}
\usage{
reset_config_recipients(
  stratification,
  configuration = "default",
  preview = FALSE
)
}
\arguments{
\item{stratification}{A vector of things you would like to stratify by. For
example, a vector of site names. But it can be anything. If you only want one
report, stratification can be equal to ""}

\item{configuration}{Which named configuration from config.yml do you want to
use? Must be "default", "prod", or "rsconnect". The default is . . . default.}

\item{preview}{Do you want to just "see" what it would look like? This is for
interactive/testing purposes. Don't use this in a script. Seriously. It won't
do anything useful. The default is zero.}
}
\value{
Returns TRUE if successful. Else it probably just dies, but it might
return FALSE.
}
\description{
Resets all to, cc, bcc entries in config.yml. The values of the
stratifications is provided by the stratification parameter.
}
