% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(emails_to_send = NULL)
}
\arguments{
\item{emails_to_send}{A table of emails to send.}
}
\value{
Not much. This is all about the side effects.
}
\description{
Sends the recipients in config.yml an email.
They get their report. We get paid. Everyone wins.
}
