% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_recipients.R
\name{validate_recipients}
\alias{validate_recipients}
\title{validate_recipients}
\usage{
validate_recipients(config)
}
\arguments{
\item{config}{A config object from config::get(). This config list must
contain 'recipients'.}
}
\value{
Nothing. If the validation fails, it will stop your script.
}
\description{
Validates that at all sites have at least one recipient. This only checks the
CURRENT active config. To test the rest of the YAML file, use validate_config.
}
