% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_to_workbook}
\alias{write_to_workbook}
\title{write_to_workbook}
\usage{
write_to_workbook(data = NULL, stratification = NA)
}
\arguments{
\item{data}{The data we are going to drop into Excel.}

\item{stratification}{Which stratification are we doing this time?}
}
\description{
Writes data to a workbook.
}
