% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_to_worksheet}
\alias{write_to_worksheet}
\title{write_to_worksheet}
\usage{
write_to_worksheet(
  wb,
  data,
  start_row = 1,
  start_col = 1,
  col_widths = "auto",
  freeze_header = TRUE,
  filter = TRUE,
  first_col_bold = FALSE,
  wrap_text = FALSE
)
}
\arguments{
\item{wb}{The workbook}

\item{data}{The data we want to export.}

\item{start_row}{The first row of the data frame.}

\item{start_col}{The first column of the data frame.}

\item{col_widths}{Either "auto" or the widths you want for your columns.
Defaults to auto.}

\item{freeze_header}{TRUE/FALSE - do you want to filter on the first row?
Defaults to TRUE.}

\item{filter}{TRUE/FALSE - do you want to add filters? Defaults to TRUE.}

\item{first_col_bold}{TRUE/FALSE - do you want to make the first column BOLD?
Defaults to TRUE.}

\item{wrap_text}{TRUE/FALSE - do you want to wrap the text? Defaults to FALSE.}
}
\description{
Writes data to a worksheet.
}
