% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import_data}
\usage{
import_data(file = "query.sql")
}
\arguments{
\item{file}{The file where the query lives.}
}
\value{
A data frame with your data.
}
\description{
Opens a SQL query, runs it, and downloads the result-set.
}
